Renamed token exceptions and updated all references
