+++ src/values/semrush/semrush-token.php
@@ -8,7 +8,7 @@
 namespace Yoast\WP\SEO\Values\SEMrush;
 
 use League\OAuth2\Client\Token\AccessTokenInterface;
-use Yoast\WP\SEO\Exceptions\SEMrush\Empty_Token_Property_Exception;
+use Yoast\WP\SEO\Exceptions\SEMrush\Tokens\Empty_Property_Exception;
 
 /**
  * Class SEMrush_Token
@@ -59,30 +59,30 @@ class SEMrush_Token {
 	 * @param bool   $has_expired   Whether or not the token has expired.
 	 * @param int    $created_at    The timestamp of when the token was created.
 	 *
-	 * @throws Empty_Token_Property_Exception Exception thrown if a token property is empty.
+	 * @throws Empty_Property_Exception Exception thrown if a token property is empty.
 	 */
 	public function __construct( $access_token, $refresh_token, $expires, $has_expired, $created_at ) {
 
 		if ( empty( $access_token ) ) {
-			throw new Empty_Token_Property_Exception( 'access_token' );
+			throw new Empty_Property_Exception( 'access_token' );
 		}
 
 		$this->access_token = $access_token;
 
 		if ( empty( $access_token ) ) {
-			throw new Empty_Token_Property_Exception( 'refresh_token' );
+			throw new Empty_Property_Exception( 'refresh_token' );
 		}
 
 		$this->refresh_token = $refresh_token;
 
 		if ( empty( $expires ) ) {
-			throw new Empty_Token_Property_Exception( 'expires' );
+			throw new Empty_Property_Exception( 'expires' );
 		}
 
 		$this->expires = $expires;
 
 		if ( is_null( $has_expired ) ) {
-			throw new Empty_Token_Property_Exception( 'has_expired' );
+			throw new Empty_Property_Exception( 'has_expired' );
 		}
 
 		$this->has_expired = $has_expired;
@@ -96,7 +96,7 @@ class SEMrush_Token {
 	 *
 	 * @return SEMrush_Token The token object.
 	 *
-	 * @throws Empty_Token_Property_Exception Exception thrown if a token property is empty.
+	 * @throws Empty_Property_Exception Exception thrown if a token property is empty.
 	 */
 	public static function from_response( AccessTokenInterface $response ) {
 		return new self(
